requirements.txt
# Computer Vision and Image Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Facial Emotion Recognition
fer==22.5.1

# Deep Learning and ML
tensorflow==2.13.0
keras==2.13.1
mtcnn==0.1.1

# Data Processing and Analysis
numpy==1.24.3
pandas==2.0.3

# Image Processing
Pillow==10.0.0

# Plotting and Visualization
matplotlib==3.7.2

# HTTP Requests
requests==2.31.0

# Additional dependencies that might be needed
scikit-learn==1.3.0
scipy==1.11.1

jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.0.0  # For interactive widgets if needed
